refactor: refactor main page and catalog page
