+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Routes, Route } from 'react-router-dom'
 import './App.css'
-import { MainPage } from './components/pages/main/mainPage'
-import { CatalogPage } from './components/pages/catalog/catalogPage'
+import { MainPage } from './components/pages/mainPage'
+import { CatalogPage } from './components/pages/catalogPage'
 import { AboutPage } from './components/pages/aboutPage'
 import { ContactsPage } from './components/pages/contacts'
