--- src/components/pages/main/mainPage.tsx
@@ -1,15 +1,0 @@
-import { Footer } from "../../Footer"
-import { Header } from "../../Header"
-import { MainConteiner } from "./mainConteiner"
-
-
-export const MainPage: React.FC = () => {
-
-    return(
-        <>
-        <Header />
-        <MainConteiner />
-        <Footer />
-        </>
-    )
-}
